+++ stack_utils.c
@@ -64,6 +64,20 @@ t_stack_node	*stack_last(t_stack_node *stk)
 	return (tmp);
 }
 
+t_stack_node	*get_cheapest(t_stack_node *a)
+{
+	t_stack_node	*tmp;
+
+	tmp = a;
+	while (tmp)
+	{
+		if (tmp->cheapest)
+			return (tmp);
+		tmp = tmp->next;
+	}
+	return (NULL);
+}
+
 int	stack_len(t_stack_node *stk)
 {
 	t_stack_node	*tmp;
@@ -77,47 +91,4 @@ int	stack_len(t_stack_node *stk)
 		tmp = tmp->next;
 	}
 	return (len);
-}
-
-void	stack_add(t_stack_node **stk, int n)
-{
-	t_stack_node	*new;
-	t_stack_node	*last_node;
-
-	if (!stk)
-		return ;
-	new = malloc(sizeof(t_stack_node));
-	if (!new)
-		print_error(*stk);
-	new->nbr = n;
-	new->next = NULL;
-	new->prev = NULL;
-	new->cheapest = false;
-	if (!*stk)
-	{
-		new->prev = NULL;
-		*stk = new;
-	}
-	else
-	{
-		last_node = stack_last(*stk);
-		last_node->next = new;
-		new->prev = last_node;
-	}
-}
-
-void free_stack(t_stack_node **stk)
-{
-    t_stack_node *current_node = *stk;
-    t_stack_node *next_node;
-
-    while (current_node != NULL)
-    {
-        next_node = current_node->next;
-        free(current_node);
-        current_node = next_node;
-    }
-
-    *stk = NULL; 
-	free(*stk);
 }
